Docs: Add Section with Solution for "double picker opening" bug (#507)
* docs: section for solve "double picker opening"

Based on the issue and the discussion, some approaches can be used. The
most common approach is described at README.

* docs: highlight the solution for an issue

This adds solution description for an issue described and commonly
searched that is not related to the lib itself, but to one of its
dependencies.
It also adds an example of usage with `<Input />` component and how to
solve it.