feat: Do not show the iOS header by default
BREAKING CHANGE: Do not show the built-in iOS header by default on iOS.
We are planning to stop the built-in iOS header support in the near future, but we'll still support custom header implementations passed to the "customHeaderIOS" prop.
For the time being, you can still show the built-in iOS header by setting the "isHeaderVisibleIOS" prop to true.