feat: Support the new iOS 14 picker style
BREAKING CHANGE: Add support for the new iOS 14 picker style by using setting the "display" prop to "inline".
You could already set "display" to "inline" to achieve an iOS 14 style, but we tweaked a bit the modal padding to look better in this use case. This is a breaking change because if you were already using the "inline" display you might have to update your style customizations to account for the new padding.
We also added a new "pickerStyle" prop that can be used to customize the picker wrapper (which is the element we use to set the padding described above).